Add HTTPS support
Now 2 servers will run. The http server runs on port 3000 (5000 for
production environments) and the https server runs on port 3001 (5001 on
in production). A bash script is included in ./bin so I don't have to
remember how to generate an SSL certificate from memory.

When making requests to the https server with curl, I found that I
needed to use the -k flag.