Add GET method to /users endpoint
The method takes a query string with the users phone number. After
testing this method, I found that I am I saving the users plain password
as well as the users hashedPassword. Right now the plain password is
being sent back on GET requests. The next commit will fix this.